Update BookPage.js
Wrote some rough draft code in the book page to fetch data from the "Get Detailed Information for a Single Book" endpoint. Still need to work on making sure it'll properly render. I am unable to test this code directly, so I'll probably have to rely on Sid or Elliot to help with that.